Update to work with 64-bit monitoring patch for 4.5

+++ zmq.hpp
@@ -496,7 +496,7 @@ namespace zmq
                 const char* data = static_cast<const char*>(zmq_msg_data(&eventMsg));
                 zmq_event_t msgEvent;
                 msgEvent.event = *(uint16_t*)data; data += sizeof(uint16_t);
-                msgEvent.value = *(int32_t*)data;
+                msgEvent.value = static_cast<uintptr_t>(*(uint64_t*)data);
                 zmq_event_t* event = &msgEvent;
 #else
                 zmq_event_t* event = static_cast<zmq_event_t*>(zmq_msg_data(&eventMsg));
